+++ module/localisation/BallLocalisation/src/BallModel.cpp
@@ -25,7 +25,6 @@
 #include "utility/math/matrix/Rotation3D.h"
 #include "utility/math/matrix/Transform2D.h"
 #include "utility/math/matrix/Transform3D.h"
-#include "utility/math/vision.h"
 #include "utility/support/eigen_armadillo.h"
 
 namespace module {
