+++ module/vision/GoalDetector/src/GoalDetector.cpp
@@ -31,7 +31,6 @@
 
 #include "utility/math/coordinates.h"
 #include "utility/math/geometry/ConvexHull.h"
-#include "utility/math/vision.h"
 #include "utility/support/eigen_armadillo.h"
 #include "utility/support/yaml_armadillo.h"
 #include "utility/vision/Vision.h"
